Cleaned up the repo
Added in a sound effect system which is basic and needs to be expanded more depending on my need, but also cleaned  up the project alot making the repo way smaller and cleaner. Decided officially to keep this skybox so got rid of all the other options as well. Over time will continmue to decrease the project size by removing what models i am not using in the project.